# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SNT Portal API
  version: 1.0.0
  description: |
    Unified API for resident/chairman portal.
servers:
  - url: http://localhost:3000
    description: local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/votes:
    get:
      summary: List votes
      security:
        - bearerAuth: []
    post:
      summary: Create vote (chairman)
      security:
        - bearerAuth: []
...